requirements.txt
python-dotenv
openai
flask